+++ hr_holidays_public/__manifest__.py
@@ -22,7 +22,6 @@
     ],
     "data": [
         "security/ir.model.access.csv",
-        "views/hr_leave_type.xml",
         "views/hr_work_location_views.xml",
         "views/hr_employee_views.xml",
         "views/menu.xml",